+++ BountyBoard.Core.Tests/BountyBoard.Core.Test.csproj
@@ -54,12 +54,13 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="AccountManagementTests.cs" />
-    <Compile Include="BasicAccountManagementTests.cs" />
+    <Compile Include="Management\AccountManagementTests.cs" />
+    <Compile Include="Management\BasicAccountManagementTests.cs" />
     <Compile Include="Extensions\AccountStuffExtensions.cs" />
+    <Compile Include="MetricsHubTests.cs" />
     <Compile Include="MockObjects\FakeRestrictiveAccess.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SeasonAchievementManagementTests.cs" />
+    <Compile Include="Management\SeasonAchievementManagementTests.cs" />
     <Compile Include="Templates.cs" />
     <Compile Include="UserRestrictedDatabaseLinkTests.cs" />
   </ItemGroup>
