+++ BountyBoard.Core/Management/AccountManagement.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BountyBoard.Core
+namespace BountyBoard.Core.Management
 {
     public class AccountManagement : UserRestrictedDatabaseLink
     {
