+++ BountyBoard.Core/Management/BasicAccountManagement.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BountyBoard.Core
+namespace BountyBoard.Core.Management
 {
     /// <summary>
     /// Basic account management for the unregistered
